perbaikan API cabang, penambahan kolom cabang di form pengeluaran pc, part 10

+++ database/migrations/2026_03_01_000004_create_pc_pengeluarans_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained()->onUpdate('cascade');
             $table->foreignId('jenis_pengeluaran_cabang_id')->constrained()->onUpdate('cascade');
             $table->date('tanggal');
-            $table->unsignedBigInteger('harga')->nullable();
+            $table->decimal('harga', 8, 2)->nullable();
             $table->string('image_lokasi', 200)->nullable();
             $table->string('image_nama', 100)->nullable();
             $table->string('image_type', 50)->nullable();
